requirements.txt
crewai>=0.80.0,<1.0.0
crewai-tools>=0.33.0,<1.0.0
aiogram>=3.7.0,<4.0.0
redis>=5.0.0,<6.0.0
rq>=1.16.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
docker>=7.0.0,<8.0.0
GitPython>=3.1.40,<4.0.0
PyGithub>=2.3.0,<3.0.0
bandit>=1.7.8,<2.0.0
litellm>=1.40.0,<2.0.0
openai>=1.30.0,<2.0.0
httpx>=0.27.0,<1.0.0
pydantic>=2.0.0,<3.0.0
